agregando controlador de usuarios y excepciones

+++ src/main/java/com/misiontic/partidosapi/controllers/EquipoController.java
@@ -2,13 +2,19 @@ package com.misiontic.partidosapi.controllers;
 
 import java.util.List;
 import java.util.Map;
+
+import javax.validation.Valid;
+
 import java.util.HashMap;
 
+import com.misiontic.partidosapi.exceptions.CustomException;
 import com.misiontic.partidosapi.models.Equipo;
 import com.misiontic.partidosapi.services.EquipoService;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.Errors;
+import org.springframework.validation.ObjectError;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -36,7 +42,10 @@ public class EquipoController {
     }
 
     @PostMapping("/equipos")
-    ResponseEntity<Map<String,String>> save(@RequestBody Equipo equipo){
+    ResponseEntity<Map<String,String>> save(@Valid @RequestBody Equipo equipo, Errors errors){
+        if(errors.hasErrors()){
+            throwError(errors);
+        }
         this._equipoService.save(equipo);
         Map<String,String> response = new HashMap<>();
         response.put("mensaje","El equipo se ha registrado correctamente");
@@ -54,10 +63,29 @@ public class EquipoController {
 
     @DeleteMapping("/equipos/{id}")
     ResponseEntity<Map<String,String>> delete(@PathVariable String id){
-        this._equipoService.deleteById(id);
+        Boolean existeEquipo = this._equipoService.existById(id);
         Map<String,String> response = new HashMap<>();
+        if(!existeEquipo){
+            response.put("mensaje","El equipo no existe en la base datos");
+            return ResponseEntity.ok(response);
+        }
+        this._equipoService.deleteById(id);
+        
         response.put("mensaje","El equipo se ha eliminado correctamente");
         return ResponseEntity.ok(response);
+    }
+///Este metodo se encarga de mostrar el error
+    public void  throwError(Errors errors){
+        String mensaje = "";
+        int index = 0;
+        for(ObjectError e:errors.getAllErrors()){
+            if(index>0){
+                mensaje+=" | ";
+            }
+            System.out.println(""+e.getDefaultMessage());
+            mensaje+=String.format("Parametro: %s - Mensaje: %s", e.getObjectName(),e.getDefaultMessage());
+        }
+        throw new CustomException(mensaje);
     }
 
 }
